Hide most of the implementation of `struct rb_io`.
